So fart so good. But I'm just gassing on.

+++ src/main/java/com/github/lehjr/mpalib/capabilities/heat/HeatStorage.java
@@ -27,27 +27,27 @@
 package com.github.lehjr.mpalib.capabilities.heat;
 
 /**
- * Based on Forge Energy and CoHF RF, but using doubles and max heat value is only a safety threshold, not a cap
+ * Based on Forge Energy and CoHF RF, but using floats and max heat value is only a safety threshold, not a cap
  */
 public class HeatStorage implements IHeatStorage {
-    protected double heat;
-    protected double capacity; // this is just a safety boundary, not an absolute cap
-    protected double maxReceive;
-    protected double maxExtract;
+    protected float heat;
+    protected float capacity; // this is just a safety boundary, not an absolute cap
+    protected float maxReceive;
+    protected float maxExtract;
 
-    public HeatStorage(double capacity) {
+    public HeatStorage(float capacity) {
         this(capacity, capacity, capacity, 0);
     }
 
-    public HeatStorage(double capacity, double maxTransfer) {
+    public HeatStorage(float capacity, float maxTransfer) {
         this(capacity, maxTransfer, maxTransfer, 0);
     }
 
-    public HeatStorage(double capacity, double maxReceive, double maxExtract) {
+    public HeatStorage(float capacity, float maxReceive, float maxExtract) {
         this(capacity, maxReceive, maxExtract, 0);
     }
 
-    public HeatStorage(double capacity, double maxReceive, double maxExtract, double heat) {
+    public HeatStorage(float capacity, float maxReceive, float maxExtract, float heat) {
         this.capacity = capacity;
         this.maxReceive = maxReceive;
         this.maxExtract = maxExtract;
@@ -55,23 +55,23 @@ public class HeatStorage implements IHeatStorage {
     }
 
     @Override
-    public double receiveHeat(double maxReceive, boolean simulate) {
+    public float receiveHeat(float maxReceive, boolean simulate) {
         if (!canReceive())
             return 0;
 
-        double heatReceived = maxReceive;
+        float heatReceived = maxReceive;
         if (!simulate)
             heat += heatReceived;
         return heatReceived;
     }
 
     @Override
-    public double extractHeat(double maxExtract, boolean simulate) {
+    public float extractHeat(float maxExtract, boolean simulate) {
         if (!canExtract()) {
             return 0;
         }
 
-        double heatExtracted = Math.min(heat, maxExtract);
+        float heatExtracted = Math.min(heat, maxExtract);
         if (!simulate) {
             heat -= heatExtracted;
         }
@@ -79,12 +79,12 @@ public class HeatStorage implements IHeatStorage {
     }
 
     @Override
-    public double getHeatStored() {
+    public float getHeatStored() {
         return heat;
     }
 
     @Override
-    public double getMaxHeatStored() {
+    public float getMaxHeatStored() {
         return capacity;
     }
