+++ src/main/java/com/github/lehjr/mpalib/capabilities/heat/IHeatStorage.java
@@ -37,7 +37,7 @@ public interface IHeatStorage {
      * @param simulate   If TRUE, the insertion will only be simulated.
      * @return Amount of heat that was (or would have been, if simulated) accepted by the storage.
      */
-    double receiveHeat(double maxReceive, boolean simulate);
+    float receiveHeat(float maxReceive, boolean simulate);
 
     /**
      * Removes heat from the storage. Returns quantity of heat that was removed.
@@ -46,17 +46,17 @@ public interface IHeatStorage {
      * @param simulate   If TRUE, the extraction will only be simulated.
      * @return Amount of heat that was (or would have been, if simulated) extracted from the storage.
      */
-    double extractHeat(double maxExtract, boolean simulate);
+    float extractHeat(float maxExtract, boolean simulate);
 
     /**
      * Returns the amount of heat currently stored.
      */
-    double getHeatStored();
+    float getHeatStored();
 
     /**
      * Returns the maximum amount of heat that can be stored.
      */
-    double getMaxHeatStored();
+    float getMaxHeatStored();
 
     /**
      * Returns if this storage can have heat extracted.
